requirements.txt
requests
Flask
Flask-Cors
Flask-Limiter
python-dotenv